feat: prepare.go is deleted
